Create LogDel_LogRetentionWorkaround
The script performs 4 main tasks: First, it removes all files with the specified file extensions (e.g., *.log, *.txt) in the given folder that are older than a certain number of days. Next, it stops the Apache2.4 service and renames all files without a number in the filename by adding a current date and "_ACL_" to the filename. Finally, the script restarts the Apache2.4 service.